+++ migrations/versions/33bb617dfb24_phase5_user_wallets.py
@@ -1,0 +1,24 @@
+"""phase5 user_wallets
+
+Revision ID: 33bb617dfb24
+Revises: 000000000001
+Create Date: 2025-09-29 22:39:57.328676
+
+"""
+
+import sqlalchemy as sa
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision = "33bb617dfb24"
+down_revision = "000000000001"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    pass
+
+
+def downgrade():
+    pass